requirements.txt
anthropic>=0.28.0
openai>=1.0.0
tqdm>=4.64.0
python-dotenv>=1.0.0
assemblyai>=0.21.0
